fix: MediaRecorderService
